fixed
Исправила замечаения
