--- src/app/interfaces/todostate-interface.ts
@@ -1,9 +1,0 @@
-import { Todo } from "./todo-interface";
-
-export interface TodoState {
-  todos: Todo[],
-}
-
-export interface AppState {
-  todos: TodoState,
-}
