--- src/app/interfaces/userstate-interface.ts
@@ -1,9 +1,0 @@
-import { User } from "./user-interface";
-
-export interface UserState {
-  users: User [];
-}
-
-export interface AppState {
-  users: UserState;
-}
