+++ src/app/todos-list/store/todo.actions.ts
@@ -4,9 +4,9 @@ import { Todo } from "../../interfaces/todo-interface"
 export const TodosActions = createActionGroup ({
   source: 'Todos',
   events: {
-  'set': props<{todos: Todo[]}>(),
-  'edit': props<{todo: Todo}>(),
-  'create': props<{todo: Todo}>(),
-  'delete': props<{id: number}>(),
+    'set': props<{todos: Todo[]}>(),
+    'edit': props<{todo: Todo}>(),
+    'create': props<{todo: Todo}>(),
+    'delete': props<{id: number}>(),
   },
 });
