+++ src/app/user-list/store/user.selector.ts
@@ -1,6 +1,13 @@
 import { createSelector } from "@ngrx/store";
-import { AppState, UserState } from "../../interfaces/userstate-interface";
+import { User } from "../../interfaces/user-interface";
 
+export interface UserState {
+  users: User [];
+}
+
+export interface AppState {
+  users: UserState;
+}
 
 export const selectUserFeature = (state: AppState) => state.users;
